+++ pyMMF/solvers/WKB.py
@@ -1,7 +1,7 @@
-'''
+"""
 Solver for parabolic graded index fibers using analytical expression of the mode profile and dispersion relation
 under the WKB approximation.
-'''
+"""
 
 import time
 import numpy as np
@@ -10,88 +10,116 @@ from joblib import Parallel, delayed
 
 from ..modes import Modes
 from ..logger import get_logger
+
 logger = get_logger(__name__)
 
 
-def solve_WKB(indexProfile, wl, **options):
-    degenerate_mode = options.get('degenerate_mode','sin')
-    n_jobs = options.get('n_jobs', -2)
-    modes = findPropagationConstants(wl,indexProfile)
-    modes = associateLPModeProfiles(modes,
-                                    wl,
-                                    indexProfile,
-                                    degenerate_mode=degenerate_mode,
-                                    n_jobs=n_jobs)
-    return modes
+def solve_WKB(indexProfile, wl, tol=1e-9, degenerate_mode="sin", n_jobs=-2):
+    r"""
+    Find the propagation constants of parabolic GRIN multimode fibers under the WKB approximation [#]_.
+    This approximation leads to inaccurate results for groups of modes close to the cutoff,
+    hence is not suitable when a limited number of modes is considered.
+    It is provided only for comparison.
+    The `IndexProfile` must be initialized with the :meth:`initParabolicGRIN
+    <pyMMF.IndexProfile.initParabolicGRIN>` method.
 
-
-def findPropagationConstants(wl, indexProfile, tol = 1e-9):
-    '''
-    Find the propagation constants of parabolic GRIN multimode fibers under the WKB approximation [1]_.
-    This approximation leads to inaccurate results for groups of modes close to the cutoff, 
-    hence is not suitable when a limited number of mode is considered.
-    It is provided only for comparison. 
-    
-    Parameters
-    ----------
-    wl : float
-    		wavelength in microns.
-        
-    indexProfile: IndexProfile object
-        object that contains data about the transverse index profile.
-        
-    Returns
+    Options
     -------
-    modes : Modes object
-        Object containing data about the modes. 
-        Note that it does not fill the transverse profiles, only the data about the propagation constants 
-        and the mode numbers.
-    
-    See Also
-    --------
-        associateLPModeProfiles()
-    
+
+    tol : float, optional
+        Tolerance on the propagation constant.
+        Default is 1e-9.
+
+    degenerate_mode : {'exp', 'sin'}, optional
+        Choice for degenerate subspaces.
+
+        - `'exp'`: returns the orbital angular momentum modes.
+          For an azimuthal index \(m > 0\), the azimuthal function is
+          \(\exp(i \cdot -m \cdot \theta)\) and \(\exp(i \cdot m \cdot \theta)\).
+
+        - `'sin'`: returns the linearly polarized modes.
+          These modes have real-valued fields with azimuthal functions
+          \(\sin(m \cdot \theta)\) and \(\cos(m \cdot \theta)\) for \(m > 0\).
+
+        Default is `'exp'`.
+
+    n_jobs : int, optional
+        Number of parallel jobs to run.
+        Default is -2, which means using all available processors.
+
     Notes
     -----
 
     Propagation constants under the WKB approximation:
 
-    .. math:: \beta_{l,m} = \sqrt{k_o^2 n_1^2-2\alpha \left( |l|+2m+1\right)}
+    .. math::
 
-    .. math:: \alpha = k_o n_1/b
+        \beta_{l,m} = \sqrt{k_o^2 n_1^2 - 2\alpha \left( |l| + 2m + 1 \right)}
 
-    with 
+    .. math::
 
-    .. math:: b = \frac{radius \times n_1}{NA}
-
-    Mode profiles under the WKB approximation
-
-    .. math:: \psi_{l,m}(r, \phi) = A e^{- \frac{\alpha r^2}{2}}(\alpha r^2)^{|m|/2} L_l^{|m|}(\alpha r^2)e^{im\phi}
+        \alpha = \frac{k_o n_1}{b}
 
     with
 
-    .. math:: L_l^{|m|}
-    
-    the Laguerre polynomials
-         
-    .. [1]  K. Okamoto, "Fundamentals of optical waveguides" 
-            Academic Press,
-            2006
-            
-    '''
+    .. math::
+
+        b = \frac{ \text{radius} \times n_1 }{ \text{NA} }
+
+    Mode profiles under the WKB approximation:
+
+    .. math::
+
+        \psi_{l,m}(r, \phi) = A e^{- \frac{\alpha r^2}{2}} (\alpha r^2)^{|m|/2} L_l^{|m|}(\alpha r^2) e^{im\phi}
+
+    where :math:`L_l^{|m|}` are the Laguerre polynomials.
+
+    .. [#] K. Okamoto, "Fundamentals of Optical Waveguides," Academic Press, 2006.
+    """
+
+    modes = findPropagationConstants(
+        wl, indexProfile, tol=tol, degenerate_mode=degenerate_mode
+    )
+    modes = associateLPModeProfiles(
+        modes, wl, indexProfile, degenerate_mode=degenerate_mode, n_jobs=n_jobs
+    )
+    return modes
+
+
+def findPropagationConstants(wl, indexProfile, tol, degenerate_mode):
+    r"""
+    Find the propagation constants of parabolic GRIN multimode fibers under the WKB approximation [#]_.
+    This approximation leads to inaccurate results for groups of modes close to the cutoff,
+    hence is not suitable when a limited number of modes is considered.
+    It is provided only for comparison.
+
+    Parameters
+    ----------
+
+        wl : float
+            wavelength in microns.
+
+        indexProfile: IndexProfile object
+            object that contains data about the transverse index profile.
+
+        tol : float, optional
+            tolerance on the propagation constant.
+            Default is 1e-9
+
+    """
     NA = indexProfile.NA
     a = indexProfile.a
-    k0 = 2.*np.pi/wl
+    k0 = 2.0 * np.pi / wl
 
     n1 = np.max(indexProfile.n)
     n2 = np.min(indexProfile.n)
 
-    b = a*n1/NA
-    
-    alpha = k0*n1/b
+    b = a * n1 / NA
 
-    beta_min = k0*n2
-    beta_max = k0*n1
+    alpha = k0 * n1 / b
+
+    beta_min = k0 * n2
+    beta_max = k0 * n1
 
     modes = Modes()
 
@@ -99,14 +127,29 @@ def findPropagationConstants(wl, indexProfile, tol = 1e-9):
     while True:
         m = 0
         while True:
-            beta_current = np.sqrt(k0**2*n1**2-2*alpha*(np.abs(m)+2*(l-1)+1))
+            beta_current = np.sqrt(
+                k0**2 * n1**2 - 2 * alpha * (np.abs(m) + 2 * (l - 1) + 1)
+            )
             if beta_current < beta_min or beta_current > beta_max:
                 break
-            degeneracy = 1 if m == 0 else 2
-            modes.betas.extend([beta_current]*degeneracy)
-            modes.number += degeneracy
-            modes.m.extend([m]*degeneracy)
-            modes.l.extend([l]*degeneracy)
+            # degeneracy = 1 if m == 0 else 2
+            # modes.betas.extend([beta_current] * degeneracy)
+            # modes.number += degeneracy
+            # modes.m.extend([m] * degeneracy)
+            # modes.l.extend([l] * degeneracy)
+            modes.betas.extend([beta_current])
+            modes.number += 1
+            modes.m.extend([m])
+            modes.l.extend([l])
+
+            if not m == 0:
+                # degenerate mode
+                modes.betas.extend([beta_current])
+                modes.number += 1
+                other_m = -m if degenerate_mode == "exp" else m
+                modes.m.extend([other_m])
+                modes.l.extend([l])
+
             m += 1
         if m == 0:
             # if we stopped with m = 0, there is no more modes
@@ -120,68 +163,83 @@ def calc_mode(modes, idx, degenerate_mode, R, TH, a, alpha):
     m = modes.m[idx]
     l = modes.l[idx]
 
-    aR2 = alpha*R.ravel()**2
+    aR2 = alpha * R.ravel() ** 2
 
     phase = m * TH.ravel()
-    psi = 0
+    # psi = 0
+
+    degenerated = False
+    if (m, l) in zip(modes.m[:idx], modes.l[:idx]):
+        degenerated = True
 
     # Non-zero transverse component
-    if degenerate_mode == 'sin':
+    if degenerate_mode == "sin":
         # two pi/2 rotated degenerate modes for m < 0
-        psi = np.pi/2 if m[idx] < 0 else 0 
-        phase_mult = np.cos(phase + psi)
+        # psi = np.pi / 2 if m < 0 else 0
+        # phase_mult = np.cos(phase + psi)
+        phase_mult = np.sin(phase) if degenerated else np.cos(phase)
 
-    elif degenerate_mode == 'exp':
+    elif degenerate_mode == "exp":
         # noticably faster than writing exp(1j*phase)
         phase_mult = np.cos(phase) + 1j * np.sin(phase)
-    
-    amplitude = np.exp(-aR2/2)*aR2**(np.abs(m)/2)*genlaguerre(l-1,np.abs(m))(aR2)
+
+    amplitude = (
+        np.exp(-aR2 / 2) * aR2 ** (np.abs(m) / 2) * genlaguerre(l - 1, np.abs(m))(aR2)
+    )
 
     Et = phase_mult * amplitude
     mode = Et.astype(np.complex64)
-    mode /= np.sqrt(np.sum(np.abs(mode)**2))
+    mode /= np.sqrt(np.sum(np.abs(mode) ** 2))
     return mode
 
-def associateLPModeProfiles(
-    modes, 
-    wl,
-    indexProfile, 
-    degenerate_mode='sin',
-    n_jobs=-2
-    ):
-    '''
+
+def associateLPModeProfiles(modes, wl, indexProfile, degenerate_mode="sin", n_jobs=-2):
+    """
     Associate the linearly polarized mode profile to the corresponding constants.
-    '''
-    
-    assert(not modes.profiles)
-    assert(degenerate_mode in ['sin', 'exp'])
+    """
+
+    assert not modes.profiles
+    assert degenerate_mode in ["sin", "exp"]
     R = indexProfile.R
     TH = indexProfile.TH
     a = indexProfile.a
     NA = indexProfile.NA
     n1 = np.max(indexProfile.n)
 
-    b = a*n1/NA
-    k0 = 2.*np.pi/wl
-    alpha = k0*n1/b
-    
-    logger.info('Finding analytical LP mode profiles associated to the propagation constants.')
+    b = a * n1 / NA
+    k0 = 2.0 * np.pi / wl
+    alpha = k0 * n1 / b
+
+    logger.info(
+        "Finding analytical LP mode profiles associated to the propagation constants."
+    )
+    logger.info(f"modes = {modes.number}")
     t0 = time.time()
 
     # for exp modes, we have +m and -m for degenerate modes
-    if degenerate_mode == 'exp':
-        modes.m = \
-            [m if not (m, l) in zip(modes.m[:idx], modes.l[:idx]) else -m for idx, (m,l) in enumerate(zip(modes.m, modes.l))]
+    if degenerate_mode == "exp":
+        modes.m = [
+            m if not (m, l) in zip(modes.m[:idx], modes.l[:idx]) else -m
+            for idx, (m, l) in enumerate(zip(modes.m, modes.l))
+        ]
 
-    R[R<np.finfo(np.float32).eps] = np.finfo(np.float32).eps
+    R[R < np.finfo(np.float32).eps] = np.finfo(np.float32).eps
 
-    
-    modes.profiles = Parallel(n_jobs = n_jobs)(
-        delayed(calc_mode)(modes, idx, degenerate_mode, R, TH, a, alpha) for idx in range(modes.number)
-        )
-    modes.sort() 
+    # modes.profiles = Parallel(n_jobs=n_jobs)(
+    #     delayed(calc_mode)(modes, idx, degenerate_mode, R, TH, a, alpha)
+    #     for idx in range(modes.number)
+    # )
+    modes.profiles = [
+        calc_mode(modes, idx, degenerate_mode, R, TH, a, alpha)
+        for idx in range(modes.number)
+    ]
+    # sort bu decreasing values of betas,
+    # in case of degenerate modes, the one with m>0 is first
+    modes.sort(lambda mode: (-mode.beta, -mode.m))
 
-    logger.info("Found %g LP mode profiles in %0.1f minutes." % (
-        modes.number, (time.time() - t0) / 60))
+    logger.info(
+        "Found %g LP mode profiles in %0.1f minutes."
+        % (modes.number, (time.time() - t0) / 60)
+    )
 
     return modes