+++ pyMMF/solvers/radial.py
@@ -8,6 +8,7 @@ Riccati's equations.
 import numpy as np
 from scipy.interpolate import interp1d
 from scipy.optimize import brentq as bisect
+from functools import partial
 from numba import jit
 import time
 import copy
@@ -23,8 +24,13 @@ CHANGE_BC_RADIUS_STEP_DEFAULT = 0.9
 N_BETA_COARSE_DEFAULT = int(1e3)
 DEFAULT_DEGENERATE_MODE = "sin"
 
+
 # choice for degenerate subspaces
-EXP_PHASE_FUNCS = [lambda x: np.exp(1j * x), lambda x: np.exp(-1j * x)]
+def _expi(x, s=1):
+    return np.exp(s * 1j * x)
+
+
+EXP_PHASE_FUNCS = [partial(_expi, s=1), partial(_expi, s=-1)]
 SIN_PHASE_FUNCS = [np.sin, np.cos]
 
 
@@ -62,8 +68,8 @@ class SmallRmaxError(Exception):
 
 
 class CalculationStopException(Exception):
-    def __init__(self):
-        self.msg = "Calculation stops"
+    def __init__(self, msg):
+        self.msg = f"Calculation stops: {msg}"
         logger.error(self.msg)
         super().__init__(self.msg)
 
@@ -111,12 +117,12 @@ def _get_field_fast(m, dh, r, nr, beta_min, delta_beta, k0):
 def get_field_fast(m, dh, r, nr, beta_min, delta_beta, k0):
     """
     Get the field calulcated using the recursive scheme for the quadratic Ricatti
-    equation as [1].
+    equation as [#]_.
 
     Be careful that numba does not allow float128, which limits the precision
     and can lead to stagnation of the recursive iterations.
 
-    .. [1] Lakshman S. Tamil, S. S. Mitra, R. Dutta, and J. M. T. Pereira,
+    .. [#] Lakshman S. Tamil, S. S. Mitra, R. Dutta, and J. M. T. Pereira,
        "Finite difference solution for graded-index cylindrical dielectric waveguides:
        a scalar wave approximation" Applied Optics, vol. 30,
        pp. 1113-1116, 1991.
@@ -146,10 +152,14 @@ def scan_betas(m, dh, r, nr, beta_min, delta_betas, k0):
 
 
 def binary_search(func, min_val, max_val, sign, beta_tol=1e-12, field_limit_tol=1e-3):
-    max_val_incr_factor = 1.1
 
     if max_val - min_val < beta_tol:
         raise PrecisionError(min_val, max_val)
+
+    if np.sign(func(min_val)) == np.sign(func(max_val)):
+        raise ValueError(
+            "Field at min_val and max_val have the same sign. Try increasing the value of r_max."
+        )
 
     converged = 0
     while not converged:
@@ -157,15 +167,10 @@ def binary_search(func, min_val, max_val, sign, beta_tol=1e-12, field_limit_tol=
             beta, binfo = bisect(
                 func, min_val, max_val, xtol=beta_tol, full_output=True
             )
-        except ValueError:
-            max_val *= max_val_incr_factor
-            logger.error(
-                f"Field at min_val={min_val} and max_val={max_val} have the same sign."
-            )
-            raise CalculationStopException(
-                "Field at min_val and max_val have the same sign."
-            )
-        converged = binfo.converged
+            converged = binfo.converged
+        except Exception as E:
+            logger.error(f"Unknown exception: {E}")
+            raise CalculationStopException(f"Unknown exception: {E}")
 
     fval = func(beta)
     if np.abs(fval) > field_limit_tol:
@@ -175,31 +180,127 @@ def binary_search(func, min_val, max_val, sign, beta_tol=1e-12, field_limit_tol=
     return beta, binfo
 
 
-def solve_radial(indexProfile, wl, **options):
+def solve_radial(
+    indexProfile,
+    wl,
+    degenerate_mode=DEFAULT_DEGENERATE_MODE,
+    min_radius_bc=MIN_RADIUS_BC_COEFF_DEFAULT,
+    change_bc_radius_step=CHANGE_BC_RADIUS_STEP_DEFAULT,
+    N_beta_coarse=N_BETA_COARSE_DEFAULT,
+    r_max=None,
+    dh=None,
+    beta_tol=None,
+    field_limit_tol=1e-3,
+    beta_min=None,
+    save_func=False,
+):
+    """
+
+    Solves the scalar wave equation for an axisymmetric index profile
+    define by a radial function.
+    The `IndexProfile` object provided must be
+    initialized with the :meth:`initFromRadialFunction
+    <pyMMF.IndexProfile.initFromRadialFunction>` method.
+
+    Options
+    -------
+    degenerate_mode : string ('exp', or 'sin'), optional
+        Choice for degenerate subspaces.
+
+        - 'exp' return the orbital angular momentum modes,
+          for an azimuthal index m>0,
+          the azimuthal function is exp(i*-m*theta) and exp(i*m*theta)
+
+        - 'sin' return the linear polarized modes,
+          they have real values of the field
+          with an azimuthal function of sin(m*theta) and cos(m*theta) for m>0.
+
+        Default is 'exp'.
+    min_radius_bc : float, optional
+        Minimum radius for the boundary condition in units of the radius of the fiber set in the index profile.
+        The algorithm will try to find a solution of the field mode profile that vanishes at this radius.
+        If not successfull, it will multiply this maximum radius by `change_bc_radius_step` and try again.
+        Default is 4.
+    change_bc_radius_step : float, optional
+        Factor to change the boundary condition radius
+        when not successfull reaching the vanishing boundary condition
+        equalt to zero at `min_radius_bc`.
+        Set the next radius to `r_max = r_max * change_bc_radius_step`.
+        Default is 0.9.
+    N_beta_coarse : int, optional
+        Number of beta values in the initial rought scan for each radial number `l`.
+        Before trying to find the beta value that satisfies the boundary condition,
+        we scan the range of betas admissible for propagating modes
+        to find the changes of the sign of the farthest radial point.
+        It gives the number of modes and initial guesses for the beta values.
+        Default is 1_000.
+    r_max : float or None, optional
+        Maximum radius for the search.
+        Default is None, it fixes the value to `np.max(indexProfile.R)`
+        i.e. the maximum radius of the index profile map.
+    dh : float or None, optional
+        Spatial resolution.
+        Note that the algorithm find the solution of a 1D problem,
+        resolution can be increased to improve the accuracy of the solution
+        with much less computational cost than the 2D problem.
+        Default is None, it fixes the value to `indexProfile.areaSize / indexProfile.npoints`.
+    beta_tol : float or None, optional
+        Tolerance for the beta value.
+        Default is  is None, it fixes the value to `np.finfo(np.float64).eps`.
+    field_limit_tol : float, optional
+        Tolerance for the field limit.
+        If the field at `r_max` is below this value, the field is considered to be zero,
+        and the boundary condition is considered to be met.
+        Default is 1e-3.
+    beta_min : float or None, optional
+        Minimum beta value considered in the search.
+        Changing this value can return non-propagating modes.
+        Default is None, it set the values to `2*np.pi/wl * n_func(r_max0)`.
+        Note that for a index profile that does not monotonically decrease with the radius,
+        this value can lead to the algorithm to miss some propagating modes.
+    save_func : bool, optional
+        Save the radial and azimuthal functions.
+        If try, it stores the results in the `data` attribute of the `Modes` object.
+        Fields are:
+        - "radial_func": the extrapolated radial function,
+        - "r_max": the maximum radius used for the search,
+        - "norm": the normalization factor of the radial function,
+        - "azimuthal_func": the azimuthal function.
+
+        Default is False.
+    """
     t0 = time.time()
 
-    degenerate_mode = options.get("degenerate_mode", DEFAULT_DEGENERATE_MODE)
+    # degenerate_mode = options.get("degenerate_mode", DEFAULT_DEGENERATE_MODE)
     phi_funcs = EXP_PHASE_FUNCS if degenerate_mode == "exp" else SIN_PHASE_FUNCS
-    min_radius_bc = options.get("min_radius_bc", MIN_RADIUS_BC_COEFF_DEFAULT)
-    change_bc_radius_step = options.get(
-        "change_bc_radius_step", CHANGE_BC_RADIUS_STEP_DEFAULT
-    )
-    N_beta_coarse = options.get("N_beta_coarse", N_BETA_COARSE_DEFAULT)
-    r_max0 = options.get("r_max", np.max(indexProfile.R))
-    dh = options.get("dh", indexProfile.areaSize / indexProfile.npoints)
-    beta_tol = options.get("beta_tol", np.finfo(np.float64).eps)
-    field_limit_tol = options.get("field_limit_tol", 1e-3)
-    save_func = options.get("save_func", False)
+    # min_radius_bc = options.get("min_radius_bc", MIN_RADIUS_BC_COEFF_DEFAULT)
+    # change_bc_radius_step = options.get(
+    #     "change_bc_radius_step", CHANGE_BC_RADIUS_STEP_DEFAULT
+    # )
+    # N_beta_coarse = options.get("N_beta_coarse", N_BETA_COARSE_DEFAULT)
+    # r_max0 = options.get("r_max", np.max(indexProfile.R))
+    dh = dh if dh is not None else indexProfile.areaSize / indexProfile.npoints
+    # dh = options.get("dh", indexProfile.areaSize / indexProfile.npoints)
+    beta_tol = beta_tol if beta_tol is not None else np.finfo(np.float64).eps
+    # beta_tol = options.get("beta_tol", np.finfo(np.float64).eps)
+    # field_limit_tol = options.get("field_limit_tol", 1e-3)
+    # save_func = options.get("save_func", False)
+    # beta_min = options.get("beta_min", None)
 
     k0 = 2.0 * np.pi / wl
 
     n_func = indexProfile.radialFunc
     radius = indexProfile.a
 
+    r_max0 = np.max(indexProfile.R) if r_max is None else r_max
+
     r = np.arange(0, r_max0 + dh, dh).astype(np.float64)
 
-    beta_min = k0 * n_func(r_max0)
+    if beta_min is None:
+        beta_min = k0 * n_func(r_max0)
     beta_max = k0 * n_func(0)
+
+    logger.info(f"Searching for modes with beta_min={beta_min}, beta_max={beta_max}")
     delta_betas = np.linspace(0, beta_max - beta_min, N_beta_coarse)
 
     modes = Modes()